+++ alembic/versions/157c0e82c08d_.py
@@ -1,0 +1,28 @@
+"""empty message
+
+Revision ID: 157c0e82c08d
+Revises: 
+Create Date: 2022-09-27 09:51:54.511911
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '157c0e82c08d'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
