+++ db/models/course.py
@@ -84,5 +84,5 @@ class CompletedContentBlock(Timestamp, Base):
     feedback = Column(Text, nullable=True)
     grade = Column(Integer, default=0)
 
-    student = relationship(User, back_populates="student_content_blocks")
-    content_block = relationship(ContentBlock, back_populates="completed_content_blocks")
+    student = relationship("User", back_populates="student_content_blocks")
+    content_block = relationship("ContentBlock", back_populates="completed_content_blocks")
